@@ -1,25 +1,21 @@
 <template>
   <div>
     <h4 class="text-gray">
-      <i class="linecons-tag" :id="transName(item)"></i>{{transName(item)}}
+      <i class="linecons-tag"></i>{{ item.name }}
     </h4>
     <div class="row">
       <div class="col-sm-3" v-for="(web, idx) in item.web" :key="idx">
-        <div class="xe-widget xe-conversations box2 label-info" title=""
-          @click="openweb(web.url)"
-          data-toggle="tooltip" 
-          data-placement="bottom" 
-          :data-original-title="web.url">
-
+        <div class="xe-widget xe-conversations box2 label-info" title="" @click="openweb(web.url)" data-toggle="tooltip"
+          data-placement="bottom" :data-original-title="web.url">
           <div class="xe-comment-entry">
             <a class="xe-user-img">
-              <img :src="web.logo" class="lozad img-circle" width="50">
+              <img :src="web.logo" class="lozad img-circle" width="40">
             </a>
             <div class="xe-comment">
-              <a href="/" class="xe-user-name overflowClip_1">
-                <strong>{{web.title}}</strong>
+              <a href="#" class="xe-user-name overflowClip_1">
+                <strong>{{ web.title }}</strong>
               </a>
-              <p class="overflowClip_2">{{web.desc}}</p>
+              <p class="overflowClip_2">{{ web.desc }}</p>
             </div>
           </div>
         </div>
@@ -29,19 +25,14 @@
   </div>
 </template>
 
-<script>
-export default {
-  name: 'WebItem',
-  props: {
-    item: Object,
-    transName: Function
-  },
-  methods: {
-    openweb(url) {
-      window.open(url, '_blank');
-    }
-  }
+<script setup lang="ts" name="WebItem">
+
+function openweb(url:any) {
+  window.open(url, '_blank');
 }
+
+const items = defineProps(['item'])
+
 </script>
 
 <style scoped>
