@@ -1,13 +1,10 @@
+<script setup lang="ts">
+</script>
+
 <template>
   <router-view></router-view>
 </template>
 
-<script>
+<style scoped>
 
-export default {
-  name: 'App'
-}
-</script>
-
-<style>
 </style>
