@@ -3,22 +3,20 @@
     <div class="footer-inner">
       <!-- 版权信息 -->
       <div class="footer-text">
-        &copy; 2018-2022
-        <a href="/"><strong>黄糖的桃花源</strong></a>
-        design by <a href="/"><strong>黄糖</strong></a>
+        &copy; 2018-2024
+        <a href="#"><strong>huangtang桃花源</strong></a> 
+        design by <a href="http://huangtang.link" target="_blank"><strong>huangtang</strong></a>
       </div>
       
       <div class="go-up">
-        <a href="/" rel="go-top"><i class="fa-angle-up"></i></a>
+        <a href="#" rel="go-top"><i class="fa-angle-up"></i></a>
       </div>
     </div>
   </footer>
 </template>
 
-<script>
-export default {
-  name: 'Footer'
-}
+<script setup lang="ts" name="Footer">
+
 </script>
 
 <style>
